fixed contributor page file structure
