+++ lib/screens/contributor.dart
@@ -4,12 +4,12 @@ import 'package:flutter/material.dart';
 import 'package:fosscu_app/constants/apikey.dart';
 import 'package:fosscu_app/constants/color.dart';
 import 'package:fosscu_app/constants/svg.dart';
-import 'package:fosscu_app/widgets/contributor_page_tile.dart';
-import 'package:fosscu_app/widgets/contributors_profile_container.dart';
-import 'package:fosscu_app/widgets/listtile.dart';
-import 'package:fosscu_app/widgets/merged_pr_container.dart';
-import 'package:fosscu_app/widgets/raised_pr_container.dart';
-import 'package:fosscu_app/widgets/top5_contributors_container.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/contributor_page_tile.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/contributors_profile_container.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/listtile.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/merged_pr_container.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/raised_pr_container.dart';
+import 'package:fosscu_app/widgets/contributor_page_widgets/top5_contributors_container.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:http/http.dart' as http;
 import 'package:shimmer_animation/shimmer_animation.dart';
